third_party/blink/web_tests/external/wpt/html/canvas/element/path-objects/2d.path.stroke.prune.arc-expected.txt

This is a testharness.js-based test.
FAIL Zero-length line segments from arcTo and arc are removed before stroking assert_equals: Red channel of the pixel at (50, 25) expected 0 but got 255
Harness: the test ran to completion.
